Refactor to useTracker
